backend/pyproject.toml
[project]
name = "anti-berojgar-backend"
version = "0.1.0"
description = "Autonomous Job Hunter Backend"
authors = [{ name = "Smeer", email = "smeer@example.com" }]
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic-settings",
    "asyncpg",
    "sqlalchemy[asyncio]",
    "langchain",
    "langchain-core",
    "langchain-google-genai",
    "langchain-groq",
    "langchain-openai",
    "langgraph",
    "langchain-mcp-adapters",
    "mcp",
    "fastmcp",
    "google-auth-oauthlib",
    "google-api-python-client",
    "python-dotenv",
    "python-multipart",
    "pillow",
    "python-jobspy",
    "supabase",
    "pypdf",
    "playwright",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agent*"]
